fix: resolve font loading issues in Storybook
