+++ package.json
@@ -49,7 +49,7 @@
     "@storybook/addon-docs": "^9.0.14",
     "@storybook/addon-onboarding": "^9.0.14",
     "@storybook/addon-themes": "^9.0.14",
-    "@storybook/nextjs-vite": "^9.0.14",
+    "@storybook/nextjs": "^9.0.17",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.6.3",
     "@testing-library/react": "^16.3.0",
