+++ src/app/fonts.ts
@@ -1,0 +1,8 @@
+import { Roboto } from 'next/font/google';
+
+export const roboto = Roboto({
+  weight: ['300', '400', '500', '700'],
+  subsets: ['latin'],
+  display: 'swap',
+  variable: '--font-roboto',
+});
