--- src/shared/theme/dark.ts
@@ -1,10 +1,0 @@
-import { createTheme } from '@mui/material/styles';
-
-export const darkTheme = createTheme({
-  palette: {
-    mode: 'dark',
-  },
-  typography: {
-    fontFamily: 'Roboto, sans-serif',
-  },
-});
