--- src/shared/theme/light.ts
@@ -1,10 +1,0 @@
-import { createTheme } from '@mui/material/styles';
-
-export const lightTheme = createTheme({
-  palette: {
-    mode: 'light',
-  },
-  typography: {
-    fontFamily: 'Roboto, sans-serif',
-  },
-});
