+++ src/shared/theme/theme.ts
@@ -1,23 +1,33 @@
 'use client';
 
-import { createTheme } from '@mui/material';
+import { createTheme } from '@mui/material/styles';
 
-import { Roboto } from 'next/font/google';
+import { roboto } from '@/app/fonts';
 
-const roboto = Roboto({
-  weight: ['300', '400', '500', '700'],
-  subsets: ['latin'],
-  display: 'swap',
-});
-
-const theme = createTheme({
-  colorSchemes: { light: true, dark: true },
-  cssVariables: {
-    colorSchemeSelector: 'class',
+export const lightTheme = createTheme({
+  palette: {
+    mode: 'light',
   },
   typography: {
     fontFamily: roboto.style.fontFamily,
   },
 });
 
-export default theme;
+export const darkTheme = createTheme({
+  palette: {
+    mode: 'dark',
+  },
+  typography: {
+    fontFamily: roboto.style.fontFamily,
+  },
+});
+
+export const theme = createTheme({
+  colorSchemes: { light: true, dark: true },
+  cssVariables: {
+    colorSchemeSelector: 'class',
+  },
+  typography: {
+    fontFamily: 'var(--font-roboto)',
+  },
+});
